almost done

+++ TopIt (1)/TopIt/src/main/java/com/example/TopIt/controllers/TeamController.java
@@ -2,15 +2,18 @@ package com.example.TopIt.controllers;
 
 import com.example.TopIt.models.Tasks;
 import com.example.TopIt.models.Teams;
+import com.example.TopIt.models.User;
 import com.example.TopIt.services.TeamsService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/team")
+@CrossOrigin(origins = "http://localhost:3000")
 public class TeamController {
     private final TeamsService teamsService;
 
@@ -36,20 +39,34 @@ public class TeamController {
     }
 
     @PostMapping("/add")
-    public ResponseEntity<Teams> addTeam(@RequestBody Teams teams) {
-        Teams newTeam= teamsService.addTeam(teams);
+    public ResponseEntity<Teams> addTeam(@RequestBody Teams teams, @AuthenticationPrincipal User u) {
+        if (!u.getAdministrator()) {
+        return new ResponseEntity<>(null, HttpStatus.valueOf(403));
+    }else {
+
+        Teams newTeam = teamsService.addTeam(teams);
         return new ResponseEntity<>(newTeam, HttpStatus.CREATED);
     }
+    }
 
-    @PutMapping("/update")
-    public ResponseEntity<Teams> updateTeam(@RequestBody Teams teams) {
-        Teams updateTeam = teamsService.updateTeam(teams);
-        return new ResponseEntity<>(updateTeam, HttpStatus.OK);
+    @PutMapping("/update/{id}")
+    public ResponseEntity<Teams> updateTeam(@RequestBody Teams teams, @AuthenticationPrincipal User u, @PathVariable Long id) {
+        if (!u.getAdministrator()) {
+            return new ResponseEntity<>(null, HttpStatus.valueOf(403));
+        }else {
+            Teams updateTeam = teamsService.updateTeam(teams,id);
+            return new ResponseEntity<>(updateTeam, HttpStatus.OK);
+        }
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> deleteTeam(@PathVariable("id") Long id) {
-        teamsService.deleteTeam(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<?> deleteTeam(@PathVariable("id") Long id, @AuthenticationPrincipal User u) {
+        if (!u.getAdministrator()) {
+            return new ResponseEntity<>(null, HttpStatus.valueOf(403));
+        } else {
+
+            teamsService.deleteTeam(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
     }
 }
